added count

+++ src/main/java/eo/org/eolang/calc/EOmul.java
@@ -15,7 +15,7 @@ public class EOmul extends EOExpressionObject {
         this.val1 = val1._setParent(this);
         this.val2 = val2._setParent(this);
         if(val1._isCalculable() && val2._isCalculable()){
-            _data = new EOData(val1._getData().toInt() * val2._getData().toInt());
+            _data = new EOData(val1._getData().toLong() * val2._getData().toLong());
         }
 
     }
@@ -25,7 +25,7 @@ public class EOmul extends EOExpressionObject {
         if(_isCalculable()){
             return _data;
         }else {
-            return new EOData(val1._getData().toInt() * val2._getData().toInt());
+            return new EOData(val1._getData().toLong() * val2._getData().toLong());
         }
     }
 }
